+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 ### Fixed
+- [NO-ISSUE] - Fix a bug where we didn't respect the overrun.
 
 ## [3.3.2] - 2023-01-08
 ### Fixed